refactor!: keep one ExperimentalJetWhaleApi for the whole project
There were two, sharing a simple name and both on a host-plugin author's
classpath — `jetwhale-annotations` reaches host-sdk through
`jetwhale-protocol:core`'s api dependency. `@OptIn(ExperimentalJetWhaleApi::class)`
therefore opted into whichever import happened to be present, silently, since
both compile. They also disagreed on retention.

The host-sdk copy is removed and its better-specified definition — BINARY
retention, an explicit message, narrowed targets — is what the shared one now
carries. The internal pair was disambiguated long ago as `InternalJetWhaleApi`
and `InternalJetWhaleHostApi`; the experimental pair never was.

BREAKING: `com.kitakkun.jetwhale.host.sdk.ExperimentalJetWhaleApi` is gone.
Host-plugin authors opting in by that name should import
`com.kitakkun.jetwhale.annotations.ExperimentalJetWhaleApi` instead, which
arrives transitively and needs no new dependency. Cheap now, expensive after 1.0.
