+++ jetwhale-host-sdk/src/main/kotlin/com/kitakkun/jetwhale/host/sdk/JetWhaleMcpCapablePlugin.kt
@@ -1,5 +1,6 @@
 package com.kitakkun.jetwhale.host.sdk
 
+import com.kitakkun.jetwhale.annotations.ExperimentalJetWhaleApi
 import kotlinx.serialization.json.JsonObject
 
 /**
