+++ jetwhale-host-sdk/src/test/kotlin/com/kitakkun/jetwhale/host/sdk/McpJsonSchemaTest.kt
@@ -2,6 +2,7 @@
 
 package com.kitakkun.jetwhale.host.sdk
 
+import com.kitakkun.jetwhale.annotations.ExperimentalJetWhaleApi
 import com.kitakkun.jetwhale.annotations.McpDescription
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.PolymorphicSerializer
