+++ jetwhale-host/core/mcp/build.gradle.kts
@@ -12,7 +12,7 @@ kotlin {
         freeCompilerArgs.add("-opt-in=com.kitakkun.jetwhale.annotations.InternalJetWhaleApi")
         // The MCP server is the consumer of the experimental plugin-facing MCP API, and host tools
         // are written with its parameter DSL, so every file in this module opts in.
-        freeCompilerArgs.add("-opt-in=com.kitakkun.jetwhale.host.sdk.ExperimentalJetWhaleApi")
+        freeCompilerArgs.add("-opt-in=com.kitakkun.jetwhale.annotations.ExperimentalJetWhaleApi")
     }
 }
